backend/requirements.txt
fastapi
uvicorn[standard]
gymnasium
gymnasium[box2d]
numpy
opencv-python-headless
motor
boto3
pillow
pymongo
motor
backoff

redis[async]